# swaggerstrayV3.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Stray",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:3000/api/v1",
			"variables": {}
		}
	],
	"paths": {
		"/admin/orders": {
			"post": {
				"tags": [
					"Order"
				],
				"summary": "createOrder",
				"description": "createOrder",
				"operationId": "createOrder",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/createOrderRequest"
									},
									{
										"example": {
											"promo": "MDK",
											"sender_addres": "fdsafasdfadsfasdM",
											"receiver_addres": "ACEsdfasdfasdfJ",
											"customer_name": "Akhmad",
											"items": [
												{
													"id_item": "1",
													"item_quantity": "5"
												}
											]
										}
									}
								]
							},
							"example": {
								"promo": "MDK",
								"sender_addres": "fdsafasdfadsfasdM",
								"receiver_addres": "ACEsdfasdfasdfJ",
								"customer_name": "Akhmad",
								"items": [
									{
										"id_item": "1",
										"item_quantity": "5"
									}
								]
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			},
			"get": {
				"tags": [
					"Order"
				],
				"summary": "getAllOrder",
				"description": "getAllOrder",
				"operationId": "getAllOrder",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/member/orders/34": {
			"patch": {
				"tags": [
					"Order"
				],
				"summary": "updateOrder",
				"description": "updateOrder",
				"operationId": "updateOrder",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/updateOrderRequest"
									},
									{
										"example": {
											"sender_addres": "Makassar ji abca",
											"status": "Paid"
										}
									}
								]
							},
							"example": {
								"sender_addres": "Makassar ji abca",
								"status": "Paid"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/member/users": {
			"get": {
				"tags": [
					"User"
				],
				"summary": "userById",
				"description": "userById",
				"operationId": "userById",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			},
			"patch": {
				"tags": [
					"User"
				],
				"summary": "updateUser",
				"description": "updateUser",
				"operationId": "updateUser",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/updateUserRequest"
									},
									{
										"example": {
											"name": "Akhmad Maknur lagi"
										}
									}
								]
							},
							"example": {
								"name": "Akhmad Maknur lagi"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
		"/admin/promos": {
			"get": {
				"tags": [
					"Promo"
				],
				"summary": "getAllPromo",
				"description": "getAllPromo",
				"operationId": "getAllPromo",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			},
			"post": {
				"tags": [
					"Promo"
				],
				"summary": "createPromo",
				"description": "createPromo",
				"operationId": "createPromo",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/createPromoRequest"
									},
									{
										"example": {
											"promo_name": "Promo Akhir Tahun",
											"promo_category": "Fashion",
											"promo_code": "FASH",
											"promo_amount": 5000
										}
									}
								]
							},
							"example": {
								"promo_name": "Promo Akhir Tahun",
								"promo_category": "Fashion",
								"promo_code": "FASH",
								"promo_amount": 5000
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
	},
	"components": {
		"schemas": {
			"createOrderRequest": {
				"title": "createOrderRequest",
				"required": [
					"promo",
					"sender_addres",
					"receiver_addres",
					"customer_name",
					"items"
				],
				"type": "object",
				"properties": {
					"promo": {
						"type": "string"
					},
					"sender_addres": {
						"type": "string"
					},
					"receiver_addres": {
						"type": "string"
					},
					"customer_name": {
						"type": "string"
					},
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Item"
						},
						"description": ""
					}
				},
				"example": {
					"promo": "MDK",
					"sender_addres": "fdsafasdfadsfasdM",
					"receiver_addres": "ACEsdfasdfasdfJ",
					"customer_name": "Akhmad",
					"items": [
						{
							"id_item": "1",
							"item_quantity": "5"
						}
					]
				}
			},
			"Item": {
				"title": "Item",
				"required": [
					"id_item",
					"item_quantity"
				],
				"type": "object",
				"properties": {
					"id_item": {
						"type": "string"
					},
					"item_quantity": {
						"type": "string"
					}
				},
				"example": {
					"id_item": "1",
					"item_quantity": "5"
				}
			},
...
			"createPromoRequest": {
				"title": "createPromoRequest",
				"required": [
					"promo_name",
					"promo_category",
					"promo_code",
					"promo_amount"
				],
				"type": "object",
				"properties": {
					"promo_name": {
						"type": "string"
					},
					"promo_category": {
						"type": "string"
					},
					"promo_code": {
						"type": "string"
					},
					"promo_amount": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"promo_name": "Promo Akhir Tahun",
					"promo_category": "Fashion",
					"promo_code": "FASH",
					"promo_amount": 5000
				}
			},
...
			"updateUserRequest": {
				"title": "updateUserRequest",
				"required": [
					"name"
				],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"example": {
					"name": "Akhmad Maknur lagi"
				}
			},
...
			"updateOrderRequest": {
				"title": "updateOrderRequest",
				"required": [
					"sender_addres",
					"status"
				],
				"type": "object",
				"properties": {
					"sender_addres": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"example": {
					"sender_addres": "Makassar ji abca",
					"status": "Paid"
				}
			},
...
		}
	},
...
}